+++ upgrade.inc.php
@@ -28,62 +28,96 @@ require_once LGLIB_PI_PATH . "/sql/{$_DB_dbms}_install.php";
 */
 function LGLIB_do_upgrade($current_ver)
 {
-    global $_LGLIB_DEFAULTS, $_LGLIB_CONF;
+    global $_LGLIB_DEFAULTS, $_LGLIB_CONF, $_PLUGIN_INFO;
 
-    $error = 0;
-
-    if ($current_ver < '0.0.2') {
-        // upgrade from 0.0.1 to 0.0.2
-        COM_errorLog("Updating Plugin to 0.0.2");
-        if (!LGLIB_do_upgrade_sql('0.0.2')) return false;
+    if (isset($_PLUGIN_INFO[$_CONF_LGLIB['pi_name']])) {
+        if (is_array($_PLUGIN_INFO[$_CONF_LGLIB['pi_name']])) {
+            // glFusion > 1.6.5
+            $current_ver = $_PLUGIN_INFO[$_CONF_LGLIB['pi_name']]['pi_version'];
+        } else {
+            // legacy
+            $current_ver = $_PLUGIN_INFO[$_CONF_LGLIB['pi_name']];
+        }
+    } else {
+        return false;
     }
 
-    if ($current_ver < '0.0.4') {
+    // Get the config object
+    $c = config::get_instance();
+
+    if (!COM_checkVersion($current_ver, '0.0.2')) {
+        // upgrade from 0.0.1 to 0.0.2
+        $current_ver = '0.0.2';
+        COM_errorLog("Updating Plugin to $current_ver");
+        if (!LGLIB_do_upgrade_sql($current_ver)) return false;
+        if (!LGLIB_do_set_version($current_ver)) return false;
+    }
+
+    if (!COM_checkVersion($current_ver, '0.0.4')) {
         // upgrade from 0.0.3 to 0.0.4
-        COM_errorLog("Updating Plugin to 0.0.4");
-        $c = config::get_instance();
+        $current_ver = '0.0.4';
+        COM_errorLog("Updating Plugin to $current_ver");
         $c->add('img_disp_relpath', $_LGLIB_DEFAULTS['img_disp_relpath'],
                 'text', 0, 0, 15, 20, true, $_LGLIB_CONF['pi_name']);
+        if (!LGLIB_do_set_version($current_ver)) return false;
     }
 
-    if ($current_ver < '0.0.5') {
+    if (!COM_checkVersion($current_ver, '0.0.5')) {
         // upgrade from 0.0.4 to 0.0.5
-        COM_errorLog("Updating Plugin to 0.0.5");
+        $current_ver = '0.0.4';
+        COM_errorLog("Updating Plugin to $current_ver");
         $c = config::get_instance();
         $c->add('cron_schedule_interval', $_LGLIB_DEFAULTS['cron_schedule_interval'],
                 'text', 0, 0, 15, 30, true, $_LGLIB_CONF['pi_name']);
         $c->add('cron_key', $_LGLIB_DEFAULTS['cron_key'],
                 'text', 0, 0, 15, 40, true, $_LGLIB_CONF['pi_name']);
+        if (!LGLIB_do_set_version($current_ver)) return false;
     }
 
-    if ($current_ver < '0.0.6') {
+    if (!COM_checkVersion($current_ver, '0.0.6')) {
         // upgrade from 0.0.5 to 0.0.6
-        COM_errorLog("Updating Plugin to 0.0.6");
+        $current_ver = '0.0.6';
+        COM_errorLog("Updating Plugin to $current_ver");
         $c = config::get_instance();
         $c->add('img_cache_interval', $_LGLIB_DEFAULTS['img_cache_interval'],
                 'text', 0, 0, 15, 50, true, $_LGLIB_CONF['pi_name']);
         $c->add('img_cache_maxage', $_LGLIB_DEFAULTS['img_cache_maxage'],
                 'text', 0, 0, 15, 60, true, $_LGLIB_CONF['pi_name']);
+        if (!LGLIB_do_set_version($current_ver)) return false;
     }
 
-    if ($current_ver < '0.0.7') {
+    if (!COM_checkVersion($current_ver, '0.0.7')) {
         // upgrade from 0.0.6 to 0.0.7
-        COM_errorLog("Updating Plugin to 0.0.7");
-        $c = config::get_instance();
+        $current_ver = '0.0.7';
+        COM_errorLog("Updating Plugin to $current_ver");
         $c->add('slimbox_autoactivation', $_LGLIB_DEFAULTS['slimbox_autoactivation'],
                 'select', 0, 0, 3, 70, true, $_LGLIB_CONF['pi_name']);
-        if (!LGLIB_do_upgrade_sql('0.0.7')) return false;
+        if (!LGLIB_do_upgrade_sql($current_ver)) return false;
+        if (!LGLIB_do_set_version($current_ver)) return false;
     }
 
-    if ($current_ver < '1.0.1') {
+    if (!COM_checkVersion($current_ver, '1.0.1')) {
         // upgrade to 1.0.1
-        COM_errorLog("Updating Plugin to 1.0.1");
-        $c = config::get_instance();
+        $current_ver = '1.0.1';
+        COM_errorLog("Updating Plugin to $current_ver");
         $c->add('use_lglib_messages', $_LGLIB_DEFAULTS['use_lglib_messages'],
                 'select', 0, 0, 3, 80, true, $_LGLIB_CONF['pi_name']);
+        if (!LGLIB_do_set_version($current_ver)) return false;
     }
 
-     return $error;
+    if (!COM_checkVersion($current_ver, '1.0.5')) {
+        // upgrade to 1.0.5
+        $current_ver = '1.0.5';
+        COM_errorLog("Updating Plugin to $current_ver");
+        $c->add('enable_smartresizer', $_LGLIB_DEFAULTS['enable_smartresizer'],
+                'select', 0, 0, 3, 90, true, $_LGLIB_CONF['pi_name']);
+        if (!LGLIB_do_set_version($current_ver)) return false;
+    }
+
+    // Final version setting in case there was no upgrade process for
+    // this version
+    if (!LGLIB_do_set_version($_LGLIB_CONF['pi_version'])) return false;
+    return true;
 }
 
 
@@ -113,6 +147,35 @@ function LGLIB_do_upgrade_sql($version)
         }
     }
     return true;
+}
+
+
+/**
+*   Update the plugin version number in the database.
+*   Called at each version upgrade to keep up to date with
+*   successful upgrades.
+*
+*   @param  string  $ver    New version to set
+*   @return boolean         True on success, False on failure
+*/
+function LGLIB_do_set_version($ver)
+{
+    global $_TABLES, $_CONF_LGLIB;
+
+    // now update the current version number.
+    $sql = "UPDATE {$_TABLES['plugins']} SET
+            pi_version = '{$_CONF_LGLIB['pi_version']}',
+            pi_gl_version = '{$_CONF_LGLIB['gl_version']}',
+            pi_homepage = '{$_CONF_LGLIB['pi_url']}'
+        WHERE pi_name = '{$_CONF_LGLIB['pi_name']}'";
+
+    $res = DB_query($sql, 1);
+    if (DB_error()) {
+        COM_errorLog("Error updating the {$_CONF_LGLIB['pi_display_name']} Plugin version",1);
+        return false;
+    } else {
+        return true;
+    }
 }
 
 ?>